+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/Adrenaline.java
@@ -54,18 +54,19 @@ public class Adrenaline extends Enchantment implements IAttributeEnchantment {
         return EnchantmentsFeature.getEfficiencyBonus(diggerItem.speed, lvl) * 2.5f * (durConsumed * durConsumed);
     }
 
-    public static float getMiningSpeedBoost(LivingEntity entity, BlockState state) {
-        ItemStack heldStack = entity.getMainHandItem();
-        if (!heldStack.isCorrectToolForDrops(state))
+    public static float getMiningSpeedBoost(ItemStack stack, LivingEntity entity, BlockState state) {
+        if (state == null || entity == null)
             return 0f;
-        int lvl = heldStack.getEnchantmentLevel(NewEnchantmentsFeature.ADRENALINE.get());
+        if (!stack.isCorrectToolForDrops(state))
+            return 0f;
+        int lvl = stack.getEnchantmentLevel(NewEnchantmentsFeature.ADRENALINE.get());
         if (lvl == 0)
             return 0f;
 
-        float miningSpeedBoost = getMiningSpeedBoost(heldStack, lvl);
+        float miningSpeedBoost = getMiningSpeedBoost(stack, lvl);
         if (miningSpeedBoost == 0f)
             return 0f;
-        return EnchantmentsFeature.applyMiningSpeedModifiers(miningSpeedBoost, false, entity);
+        return EnchantmentsFeature.applyMiningSpeedModifiers(miningSpeedBoost, state,false, entity);
     }
 
     @Override