+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/Enduring.java
@@ -1,0 +1,30 @@
+package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment;
+
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.item.enchantment.DigDurabilityEnchantment;
+import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.EnchantmentCategory;
+
+public class Enduring extends Enchantment {
+    public Enduring() {
+        super(Rarity.VERY_RARE, EnchantmentCategory.BREAKABLE, new EquipmentSlot[]{EquipmentSlot.MAINHAND});
+    }
+
+    @Override
+    public int getMinCost(int level) {
+        return 22 * level;
+    }
+
+    @Override
+    public int getMaxCost(int level) {
+        return this.getMinCost(level) + 22;
+    }
+
+    public boolean checkCompatibility(Enchantment other) {
+        return !(other instanceof DigDurabilityEnchantment) && super.checkCompatibility(other);
+    }
+
+    public static int getBonusDurability() {
+        return 500;
+    }
+}