+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/Haste.java
@@ -1,0 +1,30 @@
+package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment;
+
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.item.enchantment.DiggingEnchantment;
+import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.EnchantmentCategory;
+
+public class Haste extends Enchantment {
+    public Haste() {
+        super(Rarity.VERY_RARE, EnchantmentCategory.DIGGER, new EquipmentSlot[]{EquipmentSlot.MAINHAND});
+    }
+
+    @Override
+    public int getMinCost(int level) {
+        return 22 * level;
+    }
+
+    @Override
+    public int getMaxCost(int level) {
+        return this.getMinCost(level) + 22;
+    }
+
+    public boolean checkCompatibility(Enchantment other) {
+        return !(other instanceof DiggingEnchantment) && super.checkCompatibility(other);
+    }
+
+    public static int getBonusEfficiency() {
+        return 3;
+    }
+}